requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
asyncpg==0.31.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==5.5.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
cohere==5.20.4
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
deep-translator==1.11.4
exceptiongroup==1.3.1
fake-useragent==2.2.0
fastapi==0.128.0
fastavro==1.12.1
feedparser==6.0.12
filelock==3.19.1
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.2.4
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
langchain==0.3.27
langchain-cohere==0.4.6
langchain-community==0.3.31
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.37
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.15.0
networkx==3.2.1
nltk==3.9.2
numpy==1.26.4
orjson==3.11.5
packaging==23.2
pandas==2.2.1
pillow==10.4.0
playwright==1.57.0
playwright-stealth==2.0.1
plotly==5.19.0
portalocker==3.2.0
propcache==0.4.1
protobuf==4.25.8
psycopg2-binary==2.9.11
pyarrow==21.0.0
pydantic==2.9.2
pydantic-settings==2.11.0
pydantic_core==2.23.4
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.8.0
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2025.2
pyupbit==0.2.34
PyYAML==6.0.3
qdrant-client==1.16.1
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.27.1
sacremoses==0.1.1
safetensors==0.7.0
schedule==1.2.1
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
sentencepiece==0.2.1
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.49.3
streamlit==1.32.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.8.0
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.6
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
uuid_utils==0.14.0
uvicorn==0.39.0
websockets==12.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
